#359: made jenkins-master independently configurable

+++ zebrunner.sh
@@ -1,17 +1,39 @@
 #!/bin/bash
 
+# shellcheck disable=SC1091
+source patch/utility.sh
+
   setup() {
-    # PREREQUISITES: valid values inside ZBR_PROTOCOL, ZBR_HOSTNAME and ZBR_PORT env vars!
-    local url="$ZBR_PROTOCOL://$ZBR_HOSTNAME:$ZBR_PORT/jenkins"
+    if [[ $ZBR_INSTALLER -eq 1 ]]; then
+      # Zebrunner CE installer
+      # PREREQUISITES: valid values inside ZBR_PROTOCOL, ZBR_HOSTNAME and ZBR_PORT env vars!
+      url="$ZBR_PROTOCOL://$ZBR_HOSTNAME:$ZBR_PORT/jenkins"
+      host="$ZBR_HOSTNAME:$ZBR_PORT"
+    else
+      # load default interactive installer settings
+      # shellcheck disable=SC1091
+      source backup/settings.env.original
+
+      # load ./backup/settings.env if exist to declare ZBR* vars from previous run!
+      if [[ -f backup/settings.env ]]; then
+        source backup/settings.env
+      fi
+
+      set_jenkins_settings
+      url="$ZBR_PROTOCOL://$ZBR_HOSTNAME:$ZBR_JENKINS_PORT/jenkins"
+      host="$ZBR_HOSTNAME:$ZBR_JENKINS_PORT"
+    fi
 
     cp variables.env.original variables.env
     replace variables.env "http://localhost:8080/jenkins" "${url}"
-    replace variables.env "INFRA_HOST=localhost:8080" "INFRA_HOST=${ZBR_INFRA_HOST}"
+    replace variables.env "INFRA_HOST=localhost:8080" "INFRA_HOST=${host}"
 
     if [[ ! -z $ZBR_SONAR_URL ]]; then
       replace variables.env "SONAR_URL=" "SONAR_URL=${ZBR_SONAR_URL}"
     fi
 
+    # export all ZBR* variables to save user input
+    export_settings
   }
 
   shutdown() {
@@ -20,14 +42,27 @@
       exit 0 #no need to proceed as nothing was configured
     fi
 
+    if [ ! -f variables.env ]; then
+      echo_warning "You have to setup services in advance using: ./zebrunner.sh setup"
+      echo_telegram
+      exit -1
+    fi
+
     docker-compose --env-file .env -f docker-compose.yml down -v
     rm -f variables.env
+    rm -f backup/settings.env
   }
 
 
   start() {
     if [[ -f .disabled ]]; then
       exit 0
+    fi
+
+    if [ ! -f variables.env ]; then
+      # need proceed with setup steps in advance!
+      setup
+      exit -1
     fi
 
     # create infra network only if not exist
@@ -45,12 +80,24 @@
       exit 0
     fi
 
+    if [ ! -f variables.env ]; then
+      echo_warning "You have to setup services in advance using: ./zebrunner.sh setup"
+      echo_telegram
+      exit -1
+    fi
+
     docker-compose --env-file .env -f docker-compose.yml stop
   }
 
   down() {
     if [[ -f .disabled ]]; then
       exit 0
+    fi
+
+    if [ ! -f variables.env ]; then
+      echo_warning "You have to setup services in advance using: ./zebrunner.sh setup"
+      echo_telegram
+      exit -1
     fi
 
     docker-compose --env-file .env -f docker-compose.yml down
@@ -61,6 +108,13 @@
       exit 0
     fi
 
+    if [ ! -f variables.env ]; then
+      echo_warning "You have to setup services in advance using: ./zebrunner.sh setup"
+      echo_telegram
+      exit -1
+    fi
+
+    cp backup/settings.env backup/settings.env.bak
     cp variables.env variables.env.bak
     docker run --rm --volumes-from jenkins-master -v "$(pwd)"/backup:/var/backup "ubuntu" tar -czvf /var/backup/jenkins-master.tar.gz /var/jenkins_home
   }
@@ -70,7 +124,14 @@
       exit 0
     fi
 
+    if [ ! -f variables.env ]; then
+      echo_warning "You have to setup services in advance using: ./zebrunner.sh setup"
+      echo_telegram
+      exit -1
+    fi
+
     stop
+    cp backup/settings.env.bak backup/settings.env
     cp variables.env.bak variables.env
     docker run --rm --volumes-from jenkins-master -v "$(pwd)"/backup:/var/backup "ubuntu" bash -c "cd / && tar -xzvf /var/backup/jenkins-master.tar.gz"
     down
@@ -80,7 +141,7 @@
     if [[ -f .disabled ]]; then
       exit 0
     fi
- 
+
     source .env
     echo "jenkins-master: ${TAG_JENKINS_MASTER}"
   }
@@ -94,6 +155,41 @@
     echo "
       For more help join telegram channel: https://t.me/zebrunner
       "
+  }
+
+  set_jenkins_settings() {
+    # Setup global settings: protocol, hostname and port
+    echo "Zebrunner Jenkins General Settings"
+    local is_confirmed=0
+    if [[ -z $ZBR_HOSTNAME ]]; then
+      ZBR_HOSTNAME=`curl -s ifconfig.me`
+    fi
+
+    while [[ $is_confirmed -eq 0 ]]; do
+      read -r -p "Protocol [$ZBR_PROTOCOL]: " local_protocol
+      if [[ ! -z $local_protocol ]]; then
+        ZBR_PROTOCOL=$local_protocol
+      fi
+
+      read -r -p "Fully qualified domain name (ip) [$ZBR_HOSTNAME]: " local_hostname
+      if [[ ! -z $local_hostname ]]; then
+        ZBR_HOSTNAME=$local_hostname
+      fi
+
+      if [[ "$ZBR_PROTOCOL" == "http" ]]; then
+          ZBR_JENKINS_PORT=8080
+      else
+          ZBR_JENKINS_PORT=8443
+      fi
+
+      confirm "Zebrunner Jenkins URL: $ZBR_PROTOCOL://$ZBR_HOSTNAME:$ZBR_JENKINS_PORT/jenkins" "Continue?" "y"
+      is_confirmed=$?
+    done
+
+    export ZBR_PROTOCOL=$ZBR_PROTOCOL
+    export ZBR_HOSTNAME=$ZBR_HOSTNAME
+    export ZBR_JENKINS_PORT=$ZBR_JENKINS_PORT
+
   }
 
   echo_help() {
@@ -138,12 +234,7 @@ cd "${BASEDIR}" || exit
 
 case "$1" in
     setup)
-        if [[ $ZBR_INSTALLER -eq 1 ]]; then
           setup
-        else
-          echo_warning "Setup procedure is supported only as part of Zebrunner Server (Community Edition)!"
-          echo_telegram
-        fi
         ;;
     start)
 	start
